Fix GitHub Actions CI failures

+++ include/vkm/renderer/backend/common/render_pass.h
@@ -5,6 +5,8 @@
 #include <vkm/base/common.h>
 #include <vkm/renderer/backend/common/renderer_common.h>
 
+#include <optional>
+
 namespace vkm
 {
     constexpr const uint32_t MAX_COLOR_ATTACHMENTS = 8; // Maximum number of color attachments in a render pass
@@ -120,10 +122,10 @@ struct std::hash<vkm::VkmColorAttachmentDescriptor>
         return std::hash<uint32_t>()(descriptor._attachmentId) ^
                std::hash<uint8_t>()(static_cast<uint8_t>(descriptor._loadAction)) ^
                std::hash<uint8_t>()(static_cast<uint8_t>(descriptor._storeAction)) ^
-               std::hash<uint32_t>()(descriptor._clearColors[0]) ^
-               std::hash<uint32_t>()(descriptor._clearColors[1]) ^
-               std::hash<uint32_t>()(descriptor._clearColors[2]) ^
-               std::hash<uint32_t>()(descriptor._clearColors[3]);
+               std::hash<uint32_t>()(static_cast<uint32_t>(descriptor._clearColors[0])) ^
+               std::hash<uint32_t>()(static_cast<uint32_t>(descriptor._clearColors[1])) ^
+               std::hash<uint32_t>()(static_cast<uint32_t>(descriptor._clearColors[2])) ^
+               std::hash<uint32_t>()(static_cast<uint32_t>(descriptor._clearColors[3]));
     }
 };
 
@@ -135,7 +137,7 @@ struct std::hash<vkm::VkmDepthStencilAttachmentDescriptor>
         return std::hash<uint32_t>()(descriptor._attachmentId) ^
                std::hash<uint8_t>()(static_cast<uint8_t>(descriptor._loadAction)) ^
                std::hash<uint8_t>()(static_cast<uint8_t>(descriptor._storeAction)) ^
-               std::hash<uint32_t>()(descriptor._clearDepth) ^
+               std::hash<uint32_t>()(static_cast<uint32_t>(descriptor._clearDepth)) ^
                std::hash<uint32_t>()(descriptor._clearStencil);
     }
 };
